Clean up code
RRT is now in a class. All objects are now created in obstacles.py. robot.py is cleaned up as well. Next step is to make a main.py that has all the code we need to run a single simulation, so that the other py files are simply source files